swap change, my_randintand my_randfloat added

+++ includes/my_math.h
@@ -6,4 +6,4 @@
 double dot_product(double* a, double* b, unsigned int size);
 double my_power(double x, unsigned int n);
 int my_randInt(int minN, int maxN);
-float my_randFloat(float minN, float maxN);
+double my_randFloat(double minN, double maxN);
